project: reorganized file structure
The app/dependencies directory was renamed to di for brevity and to better describe the scope of the package. It now houses container.py and wiring.py. The providers.py file was moved to api, as this is specifically and solely to be called by routers. State was renamed to lifecycle.py, and container_keys.py was renamed to keys.py.